+++ core/src/main/java/com/dtp/core/spring/EnableDynamicTp.java
@@ -17,6 +17,6 @@ import java.lang.annotation.Target;
 @Target({ElementType.TYPE, ElementType.ANNOTATION_TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@Import(DtpBeanDefinitionRegistrar.class)
+@Import(DtpConfigurationSelector.class)
 public @interface EnableDynamicTp {
 }
