+++ starter/starter-adapter/starter-adapter-common/src/main/java/com/dtp/starter/adapter/common/autoconfigure/AdapterCommonAutoConfiguration.java
@@ -2,7 +2,7 @@ package com.dtp.starter.adapter.common.autoconfigure;
 
 import com.dtp.adapter.common.DtpAdapterListener;
 import com.dtp.common.properties.DtpProperties;
-import com.dtp.starter.common.autoconfigure.BaseBeanAutoConfiguration;
+import com.dtp.core.spring.BaseBeanAutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
