+++ starter/starter-adapter/starter-adapter-webserver/src/main/java/com/dtp/starter/adapter/webserver/autocconfigure/WebServerTpAutoConfiguration.java
@@ -3,10 +3,10 @@ package com.dtp.starter.adapter.webserver.autocconfigure;
 import com.dtp.adapter.webserver.JettyDtpAdapter;
 import com.dtp.adapter.webserver.TomcatDtpAdapter;
 import com.dtp.adapter.webserver.UndertowDtpAdapter;
+import com.dtp.core.spring.BaseBeanAutoConfiguration;
 import com.dtp.starter.adapter.webserver.autocconfigure.condition.ConditionalOnJettyWebServer;
 import com.dtp.starter.adapter.webserver.autocconfigure.condition.ConditionalOnTomcatWebServer;
 import com.dtp.starter.adapter.webserver.autocconfigure.condition.ConditionalOnUndertowWebServer;
-import com.dtp.starter.common.autoconfigure.BaseBeanAutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
