+++ starter/starter-common/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.dtp.starter.common.autoconfigure.BaseBeanAutoConfiguration
+#org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+#  com.dtp.starter.common.autoconfigure.BaseBeanAutoConfiguration
