+++ starter/starter-configcenter/cloud-starter-consul/src/main/java/com/dtp/starter/cloud/consul/autoconfigure/DtpConsulAutoConfiguration.java
@@ -1,7 +1,7 @@
 package com.dtp.starter.cloud.consul.autoconfigure;
 
 import com.dtp.starter.cloud.consul.refresh.CloudConsulRefresher;
-import com.dtp.starter.common.autoconfigure.BaseBeanAutoConfiguration;
+import com.dtp.core.spring.BaseBeanAutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
