+++ starter/starter-configcenter/cloud-starter-nacos/src/main/java/com/dtp/starter/cloud/nacos/autoconfigure/DtpCloudNacosAutoConfiguration.java
@@ -3,7 +3,7 @@ package com.dtp.starter.cloud.nacos.autoconfigure;
 import com.alibaba.cloud.nacos.NacosConfigManager;
 import com.alibaba.cloud.nacos.NacosConfigProperties;
 import com.dtp.starter.cloud.nacos.refresh.CloudNacosRefresher;
-import com.dtp.starter.common.autoconfigure.BaseBeanAutoConfiguration;
+import com.dtp.core.spring.BaseBeanAutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
