+++ starter/starter-configcenter/cloud-starter-zookeeper/src/main/java/com/dtp/starter/cloud/zookeeper/autoconfigure/DtpCloudZkAutoConfiguration.java
@@ -1,7 +1,7 @@
 package com.dtp.starter.cloud.zookeeper.autoconfigure;
 
 import com.dtp.starter.cloud.zookeeper.refresh.CloudZookeeperRefresher;
-import com.dtp.starter.common.autoconfigure.BaseBeanAutoConfiguration;
+import com.dtp.core.spring.BaseBeanAutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
